// wheretogo-backend/wheretogo-backend/src/generated/typegraphql/resolvers/inputs/PostWhereUniqueInput.ts
export class PostWhereUniqueInput {
  @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
    nullable: true
  })
  id?: number | undefined;

  @TypeGraphQL.Field(_type => [PostWhereInput], {
    nullable: true
  })
  AND?: PostWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [PostWhereInput], {
    nullable: true
  })
  OR?: PostWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [PostWhereInput], {
    nullable: true
  })
  NOT?: PostWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  authorId?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => EnumPostTypeFilter, {
    nullable: true
  })
  type?: EnumPostTypeFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  title?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  description?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableListFilter, {
    nullable: true
  })
  tags?: StringNullableListFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  likesCount?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  commentsCount?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  savedCount?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  sharesCount?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  viewsCount?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => EnumPostVisibilityFilter, {
    nullable: true
  })
  visibility?: EnumPostVisibilityFilter | undefined;

  @TypeGraphQL.Field(_type => IntNullableFilter, {
    nullable: true
  })
  locationId?: IntNullableFilter | undefined;

  @TypeGraphQL.Field(_type => IntNullableFilter, {
    nullable: true
  })
  tripId?: IntNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  createdAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  updatedAt?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeNullableFilter, {
    nullable: true
  })
  publishedAt?: DateTimeNullableFilter | undefined;

  @TypeGraphQL.Field(_type => UserRelationFilter, {
    nullable: true
  })
  author?: UserRelationFilter | undefined;

  @TypeGraphQL.Field(_type => LocationNullableRelationFilter, {
    nullable: true
  })
  location?: LocationNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => TripNullableRelationFilter, {
    nullable: true
  })
  trip?: TripNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostMediaListRelationFilter, {
    nullable: true
  })
  media?: PostMediaListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostLikeListRelationFilter, {
    nullable: true
  })
  likes?: PostLikeListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostCommentListRelationFilter, {
    nullable: true
  })
  comments?: PostCommentListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => SavedPostListRelationFilter, {
    nullable: true
  })
  saves?: SavedPostListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostShareListRelationFilter, {
    nullable: true
  })
  shares?: PostShareListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostReportListRelationFilter, {
    nullable: true
  })
  reports?: PostReportListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostCollectionItemListRelationFilter, {
    nullable: true
  })
  collections?: PostCollectionItemListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => PostMetricsNullableRelationFilter, {
    nullable: true
  })
  metrics?: PostMetricsNullableRelationFilter | undefined;
}